fix: list plans by type
