+++ application/repositories/store.go
@@ -28,7 +28,6 @@ type PlanStoreRepository interface {
 	CreatePlan(ctx context.Context, arg models.CreatePlanInput) (*models.Plan, error)
 	GetPlanByIDorSlug(ctx context.Context, idOrSlug string) (*models.Plan, error)
 	ListPlans(ctx context.Context, pagination pagination.Pagination) (*pagination.PaginationView[models.Plan], error)
-	ListPlansByType(ctx context.Context, planType models.PlanTypeEnum, pagination pagination.Pagination) (*pagination.PaginationView[models.Plan], error)
 	DeletePlanByIDorSlug(ctx context.Context, idOrSlug string) error
 	UpdatePlanByIDorSlug(ctx context.Context, idOrSlug string, arg models.UpdatePlanInput) (*models.Plan, error)
 	ArchivePlanByIDorSlug(ctx context.Context, idOrSlug string, arg models.ArchivePlanInput) error
