+++ infrastructure/postgres/store/plans/list.go
@@ -18,8 +18,8 @@ func (p *PgPlanStoreRepository) ListPlans(ctx context.Context, page pagination.P
 		Limit:      int32(page.Limit),
 		Offset:     int32(page.GetOffset()),
 		TenantSlug: ctx.Value(constants.TenantSlugKey).(string),
+		Type:       createPlanTypeEnum(page.Queries["type"]),
 	})
-
 	if err != nil {
 		p.logger.Error("Error listing plans: ", zap.Error(err))
 		return nil, postgres.MapError(err, "Postgres.ListPlans")
@@ -49,7 +49,10 @@ func (p *PgPlanStoreRepository) ListPlans(ctx context.Context, page pagination.P
 		})
 	}
 
-	count, err := p.q.CountPlans(ctx, tenantSlug)
+	count, err := p.q.CountPlans(ctx, gen.CountPlansParams{
+		TenantSlug: tenantSlug,
+		Type:       createPlanTypeEnum(page.Queries["type"]),
+	})
 	if err != nil {
 		p.logger.Error("Error counting plans: ", zap.Error(err))
 		return nil, postgres.MapError(err, "Postgres.CountMeters")
@@ -60,45 +63,9 @@ func (p *PgPlanStoreRepository) ListPlans(ctx context.Context, page pagination.P
 	return &result, nil
 }
 
-func (p *PgPlanStoreRepository) ListPlansByType(ctx context.Context, planType models.PlanTypeEnum, page pagination.Pagination) (*pagination.PaginationView[models.Plan], error) {
-	tenantSlug := ctx.Value(constants.TenantSlugKey).(string)
-	m, err := p.q.ListPlansByType(ctx, gen.ListPlansByTypeParams{
-		Type:       gen.PlanTypeEnum(planType),
-		Limit:      int32(page.Limit),
-		Offset:     int32(page.GetOffset()),
-		TenantSlug: tenantSlug,
-	})
-
-	if err != nil {
-		p.logger.Error("Error listing plans: ", zap.Error(err))
-		return nil, postgres.MapError(err, "Postgres.ListPlans")
+func createPlanTypeEnum(planType string) gen.NullPlanTypeEnum {
+	return gen.NullPlanTypeEnum{
+		PlanTypeEnum: gen.PlanTypeEnum(planType),
+		Valid:        planType != "",
 	}
-
-	plans := make([]models.Plan, 0, len(m))
-	for _, plan := range m {
-		id, err := uuid.FromBytes(plan.ID.Bytes[:])
-		if err != nil {
-			p.logger.Error("failed to parse UUID from bytes", zap.Error(err))
-			return nil, postgres.MapError(err, "Postgres.ParseUUID")
-		}
-		plans = append(plans, models.Plan{
-			Name:        plan.Name,
-			Slug:        plan.Slug,
-			Type:        models.PlanTypeEnum(plan.Type),
-			ArchivedAt:  plan.ArchivedAt,
-			Description: plan.Description.String,
-			TenantSlug:  plan.TenantSlug,
-			Base: models.Base{
-				ID:        id,
-				CreatedAt: plan.CreatedAt,
-				CreatedBy: plan.CreatedBy,
-				UpdatedBy: plan.UpdatedBy,
-				UpdatedAt: plan.UpdatedAt,
-			},
-		})
-	}
-
-	result := pagination.FormatWith(page, int(len(m)), plans)
-
-	return &result, nil
 }
