+++ interfaces/http/routes/v1/plans/list.go
@@ -13,7 +13,6 @@ import (
 
 func (h *httpHandler) list(ctx *fiber.Ctx) error {
 	tenantSlug := ctx.Get(constants.TenantHeader)
-	planType := ctx.Params("type")
 
 	// Create pagination input
 	paginationInput := pagination.ExtractPaginationFromContext(ctx)
@@ -21,22 +20,7 @@ func (h *httpHandler) list(ctx *fiber.Ctx) error {
 	var plans *pagination.PaginationView[models.Plan]
 	var err error
 
-	if planType == "" {
-		// Call service to list all plans
-		plans, err = h.planSvc.ListPlans(c, paginationInput)
-	} else {
-
-		if !models.ValidatePlanType(planType) {
-			errResp := domainerrors.NewErrorResponseWithOpts(nil, domainerrors.EINVALID, "invalid plan type")
-			h.logger.Error("invalid plan type", zap.Reflect("error", errResp))
-			return ctx.Status(errResp.Status).JSON(errResp.ToJson())
-		}
-
-		parsedPlanType := models.PlanTypeEnum(planType)
-		// Call service to list plans by type
-		plans, err = h.planSvc.ListPlansByType(c, parsedPlanType, paginationInput)
-	}
-
+	plans, err = h.planSvc.ListPlans(c, paginationInput)
 	if err != nil {
 		h.logger.Error("failed to list plans", zap.Reflect("error", err))
 		errResp := domainerrors.NewErrorResponse(err)
